on scrolling animation done correctly
