+++ src/index.css
@@ -5,3 +5,6 @@
 body {
   font-family: var(--project-font-family);
 }
+.contact {
+  height: 700px;
+}
